Enable installation of TA in CI
For each REC POD, the CD job is tied to a specific jenkins slave
that has access to the POD's lab.

For now, only UNH Ampere1 POD is added and user_config.yaml is stored
on the slave at:
  file:///data/rec_pod/user_config.yaml

Also for CD deployments, admin_user_password in user_config.yaml is
required to be the hash for "admin" to be able to pull post deployment
logs from the first controller.

Jenkins slave package requirements:
  - docker
  - jq

Co-authored-by: Cristina Pauna <cristina.pauna@enea.com>
Co-authored-by: Alexandru Avadanii <Alexandru.Avadanii@enea.com>>
Signed-off-by: Alexandru Antone <Alexandru.Antone@enea.com>
Change-Id: Iece4330c3242385c33d8a1daa092f78ce5d46cc8