validation: bluval: Use own repo logparser rules
Instead of creating the logparser rules.txt file on the fly, rely on
the project's git repository providing it.

While at it, add and configure the 'robot' publisher in order to enable
Jenkins dashboard integration and display some nice summaries directly
in the 'blueval' view (using the 'Robot Results + Duration Trend'
column), as well as on the job's overview webpage.

JIRA: VAL-104

Depends-on: I1506d84a44926e80913562f67f24885a98f94df7

Change-Id: I971ae6f8949caac8debc1aee0f35e0b35b8f65e6
Signed-off-by: Alexandru Avadanii <Alexandru.Avadanii@enea.com>